requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
loguru>=0.7.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
requests>=2.31.0
python-dotenv>=1.0.0
starlette>=0.31.1
typing-extensions>=4.8.0
jinja2>=3.1.0
MarkupSafe>=2.1.3